Adjust mouse coordinates to compensate for high DPI scaling
It seems the mouse coordinates need to be adjusted when DPI
scaling is in effect. This is logically pretty straightforward
but I can't seem to get it quite right when there's both margin
and padding. Padding isn't behaving quite the way I expect it
but this could very easily be due to my poor spatial reasoning
(I'm a little sad how long it took me to get the order of
operations right).

It sounds like Apple targets already do this scaling so it's
ifdef'd out for `__APPLE__`.